Add GitHub actions for building and testing the images
They use a reusable workflow that:

1. Sets things up
2. Builds the base image
3. Tests the base image. Right now it just runs `ckan --help` but it's a
   good indicator that everything works
4. Builds the dev image
5. Runs some logic tests that should cover most of CKAN functionality

These don't test actual endopoints of the running container as it
made things more complicated but we can explore it in the future.

To be able to run the built images locally we need the service
containers (solr, pg, etc) and to make these available to the running
container (inside the Github running container) we are exposing the
necessary ports and (ab)using the `--net=host` option.