Fix supervisor conf file owners
Although supervisor has been removed from 2.11 onwards, it is still used
in the 2.10 and 2.9 images. In #80 the supervisord.conf [was modified][1]
to run the process as the newly created `ckan-sys` user, but this same
conf file is used in the alpine image, where this user does not exist.
Conversely, the 2.9 Python image includes the same file as the alpine
one, so it doesn't use the new user.

This PR splits the files in two different versions so they use the
appropiate user.

[1]
https://github.com/ckan/ckan-docker-base/pull/80/files#diff-a9f3dc77dfb98b40d119163f7a8049572695f238c20380a6c9ca8e43c4a9daf3R4